feat: setup initial oscars integration
