+++ Cargo.toml
@@ -258,3 +258,7 @@ style = { level = "warn", priority = -1 }
 complexity = { level = "warn", priority = -1 }
 perf = { level = "warn", priority = -1 }
 pedantic = { level = "warn", priority = -1 }
+
+[patch."https://github.com/boa-dev/oscars.git"]
+oscars = { path = "../oscars/oscars" }
+oscars_derive = { path = "../oscars/oscars_derive" }
